feat: add laytime record controller

+++ tongkangku_be/Dtos/LaytimeRecord/CreateLaytimeRecordDto.cs
@@ -5,12 +5,14 @@ namespace tongkangku_be.Dtos.LaytimeRecord
     public class CreateLaytimeRecordDto
     {
         public Guid ContractId { get; set; }
+
         public OperationType OperationType { get; set; }
 
         public DateTime StartTime { get; set; }
+
         public DateTime EndTime { get; set; }
 
-        public decimal LaytimeHours { get; set; }
+        public int LaytimeHours { get; set; }
 
         public string? Notes { get; set; }
     }
